LayoutTests/fast/viewport/viewport-90.html

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE html PUBLIC "-//WAPFORUM//DTD XHTML Mobile 1.0//EN"
  "http://www.wapforum.org/DTD/xhtml-mobile10.dtd">
<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en">
<head>
    <title>XHTML Mobile Profile 1.0 document with viewport meta</title>
    <meta name="viewport" content="width=700, initial-scale=0.5">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
</body>
</html>
