Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-90-expected.txt

ALERT: viewport size 700x770 scale 0.5 with limits [0.457143, 5] and userScalable true
